Initial migration to Linux complete
Windows functionality has been abandoned in this branch.

Script has been updated to work with more up-to-date packages [`CUDA 13.0`, `pytorch 2.10.0` (nightly), `faster-whisper 1.2.0`], despite the `RealtimeSTT` PyPI package not officially supporting them.

For that to work I had to get the `python-ctranslate2` package from the AUR and edit its PKGBUILD before installing (since the default file didn't support my 3060 GPU). Then I had to create a symlink from the install location of that package in my global python install to the pyenv venv I'm using to run this script in.

Hotkeys are also enabled (for KDE Plasma 6). With some trickery they're global but only enabled when the script is running.